add to WebcamCapture can't capture to file
